adb: Fix build on big endian systems
The usb_linux_client.c file defines cpu_to_le16/32 by using the C
library htole16/32 function calls. However, cpu_to_le16/32 are used
when initializing structures, i.e in a context where a function call
is not allowed.

It works fine on little endian systems because htole16/32 are defined
by the C library as no-ops. But on big-endian systems, they are
actually doing something, which might involve calling a function,
causing build failures.

To solve this, we simply open-code cpu_to_le16/32 in a way that allows
them to be used when initializing structures.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>